backend/requirements.txt
# Core Backend Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# OpenAI & LangChain (AI Magic!)
openai==1.12.0
langchain==0.1.9
langchain-openai==0.0.5
langchain-community==0.0.20

# Vector Store (FAISS - for document search)
# Using faiss-cpu for Windows compatibility
faiss-cpu==1.7.4
tiktoken==0.6.0

# Document Processing (PDF, DOCX, etc.)
PyPDF2==3.0.1
python-docx==1.1.0
# pytesseract is optional - commenting out for easier setup
# pytesseract==0.3.10
Pillow==10.2.0

# Data & Utilities
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1

# Security (simplified for Windows)
passlib==1.7.4

# Text Processing
numpy==1.26.3
openai==1.54.0
python-dotenv==1.0.0
